# docs/api/project-settings.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Project settings — review S03/S10",
    "version": "0.10.1"
  },
  "paths": {
...
    "/api/v1/projects/{projectId}/catalogs/environments": {
      "parameters": [
        {
          "name": "projectId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          }
        }
      ],
      "get": {
        "operationId": "listEnvironments",
        "description": "",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Environments"
                  }
                }
              }
            }
          },
          "default": {
            "description": "401 phiên; 403 quyền; 404 phạm vi; 422 input; 409 version/trùng/PM cuối",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "requestId": {
                      "type": "string",
                      "maxLength": 80
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createEnvironments",
        "description": "",
        "security": [
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Environments"
                }
              }
            }
          },
          "default": {
            "description": "401 phiên; 403 quyền; 404 phạm vi; 422 input; 409 version/trùng/PM cuối",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "requestId": {
                      "type": "string",
                      "maxLength": 80
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEnvironments"
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{projectId}/catalogs/builds": {
      "parameters": [
        {
          "name": "projectId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          }
        }
      ],
      "get": {
        "operationId": "listBuilds",
        "description": "",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Builds"
                  }
                }
              }
            }
          },
          "default": {
            "description": "401 phiên; 403 quyền; 404 phạm vi; 422 input; 409 version/trùng/PM cuối",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "requestId": {
                      "type": "string",
                      "maxLength": 80
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createBuilds",
        "description": "",
        "security": [
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Builds"
                }
              }
            }
          },
          "default": {
            "description": "401 phiên; 403 quyền; 404 phạm vi; 422 input; 409 version/trùng/PM cuối",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "requestId": {
                      "type": "string",
                      "maxLength": 80
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBuilds"
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{projectId}/catalogs/devices": {
      "parameters": [
        {
          "name": "projectId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          }
        }
      ],
      "get": {
        "operationId": "listDevices",
        "description": "",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Devices"
                  }
                }
              }
            }
          },
          "default": {
            "description": "401 phiên; 403 quyền; 404 phạm vi; 422 input; 409 version/trùng/PM cuối",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "requestId": {
                      "type": "string",
                      "maxLength": 80
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createDevices",
        "description": "",
        "security": [
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Devices"
                }
              }
            }
          },
          "default": {
            "description": "401 phiên; 403 quyền; 404 phạm vi; 422 input; 409 version/trùng/PM cuối",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "requestId": {
                      "type": "string",
                      "maxLength": 80
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDevices"
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{projectId}/catalogs/categories": {
      "parameters": [
        {
          "name": "projectId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          }
        }
      ],
      "get": {
        "operationId": "listCategories",
        "description": "",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Categories"
                  }
                }
              }
            }
          },
          "default": {
            "description": "401 phiên; 403 quyền; 404 phạm vi; 422 input; 409 version/trùng/PM cuối",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "requestId": {
                      "type": "string",
                      "maxLength": 80
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createCategories",
        "description": "",
        "security": [
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Categories"
                }
              }
            }
          },
          "default": {
            "description": "401 phiên; 403 quyền; 404 phạm vi; 422 input; 409 version/trùng/PM cuối",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "requestId": {
                      "type": "string",
                      "maxLength": 80
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategories"
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{projectId}/catalogs/milestones": {
      "parameters": [
        {
          "name": "projectId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          }
        }
      ],
      "get": {
        "operationId": "listMilestones",
        "description": "",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Milestones"
                  }
                }
              }
            }
          },
          "default": {
            "description": "401 phiên; 403 quyền; 404 phạm vi; 422 input; 409 version/trùng/PM cuối",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "requestId": {
                      "type": "string",
                      "maxLength": 80
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createMilestones",
        "description": "",
        "security": [
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Milestones"
                }
              }
            }
          },
          "default": {
            "description": "401 phiên; 403 quyền; 404 phạm vi; 422 input; 409 version/trùng/PM cuối",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "requestId": {
                      "type": "string",
                      "maxLength": 80
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMilestones"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "TMS_SESSION"
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-TOKEN"
      }
    },
    "schemas": {
...
      "Environments": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "projectId": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "code": {
            "type": "string",
            "maxLength": 32
          },
          "name": {
            "type": "string",
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "nullable": true
          },
          "version": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "active": {
            "type": "boolean"
          }
        },
        "required": []
      },
      "CreateEnvironments": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 32
          },
          "name": {
            "type": "string",
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "nullable": true
          }
        },
        "required": [
          "code",
          "name"
        ]
      },
...
      "Builds": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "projectId": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "platform": {
            "type": "string",
            "maxLength": 32
          },
          "versionLabel": {
            "type": "string",
            "maxLength": 50
          },
          "buildNumber": {
            "type": "string",
            "maxLength": 50,
            "nullable": true
          },
          "notes": {
            "type": "string",
            "maxLength": 500,
            "nullable": true
          },
          "releasedAt": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "version": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        },
        "required": []
      },
      "CreateBuilds": {
        "type": "object",
        "properties": {
          "platform": {
            "type": "string",
            "maxLength": 32
          },
          "versionLabel": {
            "type": "string",
            "maxLength": 50
          },
          "buildNumber": {
            "type": "string",
            "maxLength": 50,
            "nullable": true
          },
          "notes": {
            "type": "string",
            "maxLength": 500,
            "nullable": true
          },
          "releasedAt": {
            "type": "string",
            "format": "date",
            "nullable": true
          }
        },
        "required": [
          "platform",
          "versionLabel"
        ]
      },
...
      "Devices": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "projectId": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "code": {
            "type": "string",
            "maxLength": 32
          },
          "name": {
            "type": "string",
            "maxLength": 100
          },
          "model": {
            "type": "string",
            "maxLength": 100,
            "nullable": true
          },
          "osName": {
            "type": "string",
            "maxLength": 50,
            "nullable": true
          },
          "osVersion": {
            "type": "string",
            "maxLength": 50,
            "nullable": true
          },
          "version": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "active": {
            "type": "boolean"
          }
        },
        "required": []
      },
      "CreateDevices": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 32
          },
          "name": {
            "type": "string",
            "maxLength": 100
          },
          "model": {
            "type": "string",
            "maxLength": 100,
            "nullable": true
          },
          "osName": {
            "type": "string",
            "maxLength": 50,
            "nullable": true
          },
          "osVersion": {
            "type": "string",
            "maxLength": 50,
            "nullable": true
          }
        },
        "required": [
          "code",
          "name"
        ]
      },
...
      "Categories": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "projectId": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "code": {
            "type": "string",
            "maxLength": 32
          },
          "name": {
            "type": "string",
            "maxLength": 100
          },
          "version": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "active": {
            "type": "boolean"
          }
        },
        "required": []
      },
      "CreateCategories": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 32
          },
          "name": {
            "type": "string",
            "maxLength": 100
          }
        },
        "required": [
          "code",
          "name"
        ]
      },
...
      "Milestones": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "projectId": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "code": {
            "type": "string",
            "maxLength": 32
          },
          "name": {
            "type": "string",
            "maxLength": 100
          },
          "startsOn": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "dueOn": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "version": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "archived": {
            "type": "boolean"
          }
        },
        "required": []
      },
      "CreateMilestones": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 32
          },
          "name": {
            "type": "string",
            "maxLength": 100
          },
          "startsOn": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "dueOn": {
            "type": "string",
            "format": "date",
            "nullable": true
          }
        },
        "required": [
          "code",
          "name"
        ]
      },
...
    }
  }
}